+++ packages/node-utils/src/Security.ts
@@ -3,6 +3,18 @@ import * as crypto from 'crypto'
 import { Hashes } from './Hashes'
 
 const defaultAlgorithm = 'aes-256-cbc'
+
+const charsets = {
+  NUMBERS: '0123456789',
+  LOWERCASE: 'abcdefghijklmnopqrstuvwxyz',
+  UPPERCASE: 'ABCDEFGHIJKLMNOPQRSTUVWXYZ',
+  SYMBOLS: '!#$%&()+,-.<=>?@^_'
+}
+
+export interface GeneratedKeyPair {
+  publicKey: string
+  privateKey: string
+}
 class SecurityClass {
   getCryptoParams(secretKey: string): { key: string; iv: string } {
     const key = Hashes.md5(secretKey)
@@ -75,6 +87,96 @@ class SecurityClass {
     const decipher = crypto.createDecipheriv(defaultAlgorithm, key, iv)
 
     return Buffer.concat([decipher.update(payload), decipher.final()])
+  }
+
+  async generateElipticCurve(): Promise<GeneratedKeyPair> {
+    const keys = await new Promise<GeneratedKeyPair>((resolve, reject) => {
+      try {
+        crypto.generateKeyPair(
+          'ed25519',
+          {
+            publicKeyEncoding: {
+              type: 'spki',
+              format: 'pem'
+            },
+            privateKeyEncoding: {
+              type: 'pkcs8',
+              format: 'pem'
+            }
+          },
+          (error, publicKey, privateKey) => {
+            if (error) {
+              reject(error)
+              return
+            }
+
+            resolve({ publicKey, privateKey })
+          }
+        )
+      } catch (error) {
+        reject(error)
+        return
+      }
+    })
+
+    return {
+      privateKey: keys.privateKey,
+      publicKey: keys.publicKey
+    }
+  }
+
+  /**
+   * Encrypts a message using an elliptic curve privateKey
+   * Returns a Base64 string with the generated encryption
+   * @param message
+   * @param privateKey
+   * @returns
+   */
+  encryptStringWithElliptic(message: string, privateKey: string): string {
+    const encrypted = crypto.sign(null, Buffer.from(message), privateKey)
+
+    return encrypted.toString('base64')
+  }
+
+  /**
+   * Verifies if the message is equivalent to the encryptedBase64
+   * using an Elliptic Public Key
+   * @param message String
+   * @param encryptedMessageBase64 Elliptic key encrypted Base64
+   * @param publicKey Elliptic Public Key
+   * @returns
+   */
+  verifySignedStringWithElliptic(
+    message: string,
+    encryptedMessageBase64: string,
+    publicKey: string
+  ): boolean {
+    const verified = crypto.verify(
+      null,
+      Buffer.from(message),
+      publicKey,
+      Buffer.from(encryptedMessageBase64, 'base64')
+    )
+
+    return verified
+  }
+
+  generatePassword(length: number): string {
+    const charset =
+      charsets.LOWERCASE +
+      charsets.UPPERCASE +
+      charsets.NUMBERS +
+      charsets.SYMBOLS
+
+    const charsetLength = charset.length
+
+    let password = ''
+
+    while (length--) {
+      password += charset[crypto.randomInt(charsetLength)]
+    }
+
+    return password
   }
 }